+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Changed:
 * use [json-stable-stringify](https://github.com/substack/json-stable-stringify)
   to convert objects into strings.
 
+Fixed:
+
+* make `id` in `server#addOne` optional
+
 
 ## 0.1.0 - 17/02/2016